requirements.txt
absl-py==2.4.0
alabaster==1.0.0
altair==6.0.0
annotated-doc==0.0.4
appnope==0.1.4
asttokens==3.0.1
astunparse==1.6.3
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
black==26.3.1
blinker==1.9.0
bs4==0.0.2
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
comm==0.2.3
contourpy==1.3.2
coverage==7.13.5
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
docutils==0.21.2
exceptiongroup==1.3.1
executing==2.2.1
fatsecret==0.5.0
filelock==3.25.2
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.3.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.80.0
h5py==3.14.0
idna==3.7
imagesize==2.0.0
iniconfig==2.3.0
ipykernel==7.2.0
ipython==8.39.0
isort==8.0.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
keras==3.12.1
kiwisolver==1.5.0
libclang==18.1.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.19.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.13.0.92
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==2.3.3
parso==0.8.6
pathspec==1.0.4
pexpect==4.9.0
pi_heif==1.3.0
pillow==12.2.0
pillow-avif-plugin==1.5.5
platformdirs==4.9.4
pluggy==1.6.0
polars==1.39.3
polars-runtime-32==1.39.3
prompt_toolkit==3.0.52
protobuf==7.34.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pydeck==0.9.1
Pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.3
pytest-cov==7.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
rauth==0.7.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
roboflow==1.3.1
rpds-py==0.30.0
ruff==0.15.9
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
snowballstemmer==3.0.1
soupsieve==2.8.3
Sphinx==8.1.3
sphinx_rtd_theme==3.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
tensorflow==2.21.0
termcolor==3.3.0
toml==0.10.2
tomli==2.4.1
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing_extensions==4.15.0
tzdata==2026.1
ultralytics==8.4.34
ultralytics-thop==2.0.18
urllib3==2.6.3
wcwidth==0.6.0
wrapt==2.1.2
